Fix Gtk3.21 segfault using GtkStyleContext
In GTK 3.21, the msd_osd_window_style_set segfaults on GtkStyle. This really should be updated for all GTK3 versions, so port it to GtkStyleContext. Just this one function for now as it stops the OSD crashes. Test built over GTK 3.21, no apparent problems and ran fine on GTK 3.18 as well, no apparent differences in theming using any theme from 3.18 with master.