Request keyboard event reports from global keybindings of other windows
This allows mate-settings-daemon to honor global keybindings whenever
a different application has keybindings that seem conflicting, but are
usually not.

For example, an application defining a `Super_L` keybinding
and MATE having defining a shortcut that uses the `<Mod4>` key. They are
the same physical keyboard, but one is a modifier and the other is not.
Without requesting keyboard reports from other windows, the `<Mod4>`
shortcuts are swallowed whenever there is a `Super_L` shortcut
elsewhere.