Prevent media-keys bindings from swallowing modifier keys
This changes how the media-keys manager determines successful
key-bindings from KeyRelease to KeyPress so that it does not conflict
with third-party applications that want to use modifier keys as global
key-bindings. We've already done this change for general keybindings and
marco.

It also cleans up a bunch of deprectation warnings.