Add setting for adjustment of audio volume above 100 per cent:  Part 3
There is often a need for the user to increase the audio playback volume above
the volume level known as "100% volume".  While increasing the audio volume
above 100% can result in degraded audio quality, sometimes the audio was, for
example, originally recorded at an extremely low volume, and the user has no
other option to clearly hear the audio.  Unfortunately, most MATE applications
with volume controls do not allow the user to set the volume level above 100%.
For example, the main MATE Sound Preferences dialog lets you set the audio
volume beyond 100% (when possible), whereas the Volume Control Applet, Volume
Control status icon, and special "multimedia" volume control keys do not.  In
fact, if the user even tries to change the volume using any of the latter
methods, and the current volume level is above 100%, these latter methods will
all reduce the volume to 100%, even if the user tried to increase the volume!

This is part 3 of a patch to change this situation.  This patch adds this
capability to the handlers for the "multimedia" volume control keys -- if the
appropriate setting is enabled in the MATE Volume Control Dialog (see
patch 2), then the user can increase the audio volume beyond 100% by pressing
the "Volume Up" key on their keyboard (if they have such a key).  While this
patch is smaller than patch 2, it is equally important since the original
feature request was for the multimedia keys and not for anything else in
particular.