configure.ac: don't link plugins with -no-undefined
Many of the plugins depend on mate_settings_plugin_get_type(), but this
symbol is only available at runtime in the mate-settings-daemon program.

With GNU libtool the -no-undefined flag gets silently ignored and the
build succeeds, but with slibtool it correctly sets -Wl,--no-undefined
in the LDFLAGS causing the build to fail.

To solve this simply removing -no-undefined is enough.

Gentoo issue: https://bugs.gentoo.org/922326